Add modules. Extract functions to modules
